server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.6.2",
    "dnspython>=2.8.0",
    "eventlet>=0.40.4",
    "facenet-pytorch>=2.5.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.5",
    "motor>=3.7.1",
    "mtcnn>=1.0.0",
    "opencv-python>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "redis>=7.1.1",
    "torch>=2.10.0",
    "argon2-cffi>=23.1.0",
]
